int64 chatID fix

+++ src/CTBot.h
@@ -105,8 +105,8 @@ public:
 	//             (in json format or using the CTBotInlineKeyboard class helper)
 	// returns
 	//   true if no error occurred
-	bool sendMessage(int32_t id, String message, String keyboard = "");
-	bool sendMessage(int32_t id, String message, CTBotInlineKeyboard &keyboard);
+	bool sendMessage(int64_t id, String message, String keyboard = "");
+	bool sendMessage(int64_t id, String message, CTBotInlineKeyboard &keyboard);
 
 	// terminate a query started by pressing an inlineKeyboard button. The steps are:
 	// 1) send a message with an inline keyboard
